Rename SelectionSessionModule to LegacySelectionSessionModule
